New Features
The latest version of the Jarvis project, available on Github, includes a new Jarvis class with several exciting features. This virtual assistant can perform a variety of tasks, including website lookups, playing YouTube videos, retrieving stock market information, creating art, and executing a morning protocol.

To activate the morning protocol, simply say "activate morning protocol" to Jarvis, and he will launch a website of your choice, such as a news or weather site. This feature is customizable to suit your needs, allowing you to select the website that is most useful to you in the morning.